Move espr::parser::{Statement, Function, Procedure} to espr::ast::algorithm::
