+++ espr/src/ast/mod.rs
@@ -1,4 +1,5 @@
 //! Abstract Syntax Tree (AST) of EXPRESS Language
 
+pub mod algorithm;
 pub mod expression;
 pub mod types;
